+++ tools/tidy.sh
@@ -19,11 +19,14 @@ x() {
     if [[ -n "${verbose:-}" ]]; then
         (
             set -x
-            "$cmd" "$@"
+            "${cmd}" "$@"
         )
     else
-        "$cmd" "$@"
+        "${cmd}" "$@"
     fi
+}
+warn() {
+    echo >&2 "warning: $*"
 }
 
 if [[ "${1:-}" == "-v" ]]; then
@@ -46,16 +49,16 @@ fi
 if type -P shfmt &>/dev/null; then
     x shfmt -l -w $(git ls-files '*.sh')
 else
-    echo >&2 "WARNING: 'shfmt' is not installed"
+    warn "'shfmt' is not installed"
 fi
 if type -P "${prettier}" &>/dev/null; then
     x "${prettier}" -l -w $(git ls-files '*.yml')
     x "${prettier}" -l -w $(git ls-files '*.js')
 else
-    echo >&2 "WARNING: 'prettier' is not installed"
+    warn "'prettier' is not installed"
 fi
 if type -P shellcheck &>/dev/null; then
     x shellcheck $(git ls-files '*.sh')
 else
-    echo >&2 "WARNING: 'shellcheck' is not installed"
+    warn "'shellcheck' is not installed"
 fi